refactor(managers): 抽离分发内核到 base.py，消除 5 个门面的重复代码
将 downloader/mediaserver/notification/storage/mediarecognize 五个门面管理器中
逐字重复的「分发内核」（单例、空值判定、系统/限流/插件错误处理、系统后端面 + 插件
钩子面、同步/异步两步分发）收敛到 app/managers/base.py 的三层结构：

- ManagerBase（Singleton 元类 + 系统后端面 + 单面 _dispatch）
- PluginDispatchManager（+ 插件钩子面 + 两步 dispatch）
- AsyncDispatchMixin（异步 async_dispatch，仅 mediarecognize 用）

五个子类改为继承对应基类，仅保留各自类型化对外方法，净删除 857 行重复。
行为保持：去除 group B「同一 traceback 重复记录」的日志冗余，统一 group A/B 的
错误处理与限流日志命名。facade 68 + 广域 manager 域 99 测试全绿。
